+++ bin/iron/Cargo.toml
@@ -1,7 +1,29 @@
 [package]
 name = "iron"
-version.workspace = true
+version = "0.4.1"
 edition.workspace = true
+license-file.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[build-dependencies]
+tauri-build = { version = "1.2", features = [] }
 
 [dependencies]
 iron-core = { path = "../../crates/core" }
+tauri = { version = "1.2", features = ["api-all", "system-tray"] }
+tokio = { version = "1.26.0", features = ["full", "sync"] }
+fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs", rev = "1a4c224" }
+tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+thiserror = "1.0.40"
+color-eyre = "0.6.2"
+env_logger = "0.10.0"
+eyre = "0.6.8"
+once_cell = "1.17.1"
+
+[features]
+default = []
+# this feature is used for production builds or when `devPath` points to the filesystem
+# DO NOT REMOVE!!
+custom-protocol = ["tauri/custom-protocol"]
+debug = ["tauri/devtools"]
