+++ bin/iron/src/error.rs
@@ -1,0 +1,24 @@
+use crate::app;
+
+#[derive(thiserror::Error, Debug)]
+pub enum AppError {
+    #[error(transparent)]
+    Core(#[from] iron_core::error::Error),
+
+    #[error(transparent)]
+    DB(#[from] iron_core::db::Error),
+
+    #[error(transparent)]
+    FixPathEnv(#[from] fix_path_env::Error),
+
+    #[error(transparent)]
+    Eyre(#[from] color_eyre::eyre::Error),
+
+    #[error(transparent)]
+    WindowSend(#[from] tokio::sync::mpsc::error::SendError<iron_core::app::Event>),
+
+    #[error(transparent)]
+    TauriError(#[from] tauri::Error),
+}
+
+pub type Result<T> = std::result::Result<T, AppError>;
