+++ crates/core/src/lib.rs
@@ -1,0 +1,16 @@
+pub mod abis;
+pub mod alchemy;
+pub mod app;
+pub mod block_listener;
+pub mod crypto;
+pub mod db;
+pub mod dialogs;
+pub mod error;
+pub mod foundry;
+pub mod networks;
+pub mod peers;
+pub mod rpc;
+pub mod settings;
+pub mod types;
+pub mod wallets;
+pub mod ws;
